moved samples to module, removing NLO specific objects
