+++ autofit/tools/text_util.py
