+++ test_autofit/unit/optimize/test_output.py
@@ -2,13 +2,13 @@ import pytest
 
 import autofit as af
 from autofit import Paths
-from autofit.optimize.non_linear.output import AbstractOutput
+from autofit.optimize.non_linear.samples import AbstractSamples
 from test_autofit.mock import MockClassNLOx4, MockClassNLOx6
 
 pytestmark = pytest.mark.filterwarnings("ignore::FutureWarning")
 
 
-class MockOutput(AbstractOutput):
+class MockOutput(AbstractSamples):
     def __init__(
         self,
         model,
